requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
appdirs==1.4.4
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
coverage==7.6.12
cssselect==1.2.0
cycler==0.12.1
distro==1.9.0
fake-useragent==2.0.3
Flask==3.1.0
fonttools==4.56.0
fredapi==0.5.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
kiwisolver==1.4.8
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.3
openai==1.65.4
packaging==24.2
pandas==2.2.3
parse==1.20.2
pillow==11.1.0
pluggy==1.5.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyee==11.1.1
pyparsing==3.2.1
pyppeteer==2.0.0
pyquery==2.0.1
pytest==8.3.2
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
requests==2.32.3
requests-html==0.10.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
w3lib==2.3.1
websockets==10.4
Werkzeug==3.1.3
zipp==3.21.0
